Refs 9671. Separate pool and dhared history
The Payload pool is still in shared memory, but the reader
will be notified about new entries on the 'history' shared buffer
that holds offsets to the elements in the pool.
This way the writer can manage the payloads in the pool
in arbitrary order, allowing for the user
to request and release payloads or loans whenever it needs.
The only requirement is that payloads that enter to the history
are removed fro the history in the same order, i.e.,
payload for sequence number 7 cannot be removed before
payload for sequence number 5.

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>