Refs 9919. Cycle through datasharing writers faster
Instead of staying on a writer until all its payloads have been
processed, take a snapshot of the last payload at the moment of
starting the processing and process just until there. If the same writer
has added more payloads while we were processing, those will be taken on
the next iteration.

This is to avoid one very fast writer blocking the processing forever,
as the listener may not be able to reach the end of published payloads.

Signed-off-by: Iker Luengo <ikerluengo@eprosima.com>